clientes/views.py
from clientes.models import Cliente
from clientes.serializer import ClientesSerializer
from rest_framework import viewsets

class ClienteViewSet(viewsets.ModelViewSet):
    queryset = Cliente.objects.all()
    serializer_class = ClientesSerializer
